spark/notebooks/lib/santiagomdeo/producerStreaming/requirements.txt
requests
beautifulsoup4
kafka-python
